% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bias_info}
\alias{bias_info}
\title{Bias table information}
\format{A data frame with 30 rows and 3 variables:
\describe{
  \item{parameter}{parameter name}
  \item{description}{description of the parameter}
  \item{unit}{unit}
}}
\usage{
bias_info
}
\description{
A dataset containing the parameters order and description.
}
\keyword{datasets}
