% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q_ss.R
\name{C_init}
\alias{C_init}
\title{Initializing the steady states of the decomposing pools}
\usage{
C_init(lat, site, species, climate, thinning, parameters, parsQlitter, soil,
  size_dist, settings = list(light_model = 2, transp_model = 2, phys_model =
  2, height_model = 1, correct_bias = 0, calculate_d13c = 0))
}
\arguments{
\item{settings}{}

\item{latitude}{self explanatory}
}
\description{
Initializing the steady states of the decomposing pools
}
