% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiomassPartition.R
\name{calc_biomass_partition}
\alias{calc_biomass_partition}
\title{Calculations for partioning biomass}
\usage{
calc_biomass_partition(NPP, avDBH, modifier_physiology, m0, FR, pfsConst,
  pfsPower, pRx, pRn)
}
\description{
Internal functions for partitioning biomass
}
