% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StableIsotopes.R
\name{calc_d13c}
\alias{calc_d13c}
\title{Calculations for stable isotopes}
\usage{
calc_d13c(T_av, CaMonthly, D13Catm, elev, GPPmolc, month, canopy_conductance,
  RGcGW, D13CTissueDif, aFracDiffu, bFracRubi)
}
\description{
Internal functions for isotope calculations
}
