% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaterBalance.R
\name{calc_transpiration_PM}
\alias{calc_transpiration_PM}
\title{Internal calculations Water Balance}
\usage{
calc_transpiration_PM(Q, VPD, h, gBL, gC)
}
\description{
Use Penman-Monteith equation for computing canopy transpiration
}
