% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climate_eum}
\alias{climate_eum}
\title{Climate forcing data}
\format{A data frame with 156 rows and 7 variables:
\describe{
  \item{year}{calendar year}
  \item{month}{month}
  \item{tmp_min}{minimum monthly temperature}
  \item{tmp_max}{maximum monthly temperature}
  \item{tmp_ave}{average monthly temperature}
  \item{prcp}{total monthly precipittation}
  \item{srad}{solar radiation}
  \item{frost_days}{number of frost days}
  \item{co2}{atmospheric co2}
  \item{d13catm}{d13catm}
}}
\usage{
climate_eum
}
\description{
Monthly values for climatic data
}
\keyword{datasets}
