% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_climate}
\alias{d_climate}
\title{Climate input}
\format{A \code{data frame} with 156 rows and 7 variables:
\describe{
  \item{year}{calendar year}
  \item{month}{month}
  \item{tmp_min}{monthly mean daily minimum temperature (C)}
  \item{tmp_max}{monthly mean daily maximum temperature (C)}
  \item{tmp_ave}{monthly mean daily maximum temperature (C). (optional)}
  \item{prcp}{monthly rainfall (mm month-1)}
  \item{srad}{monthly mean daily solar radiation (MJ m-2 d-1)}
  \item{frost_days}{frost days per month (d month-1)}
  \item{co2}{monthly mean atmospheric co2 (ppm), required if calculate_d13c=1 (optional)}
  \item{d13catm}{Monthly mean isotopic composition of air (‰), required if calculate_d13c=1 (optional)}
}}
\usage{
d_climate
}
\description{
Table containing the information about monthly values for climatic data.
}
\keyword{datasets}
