% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_defoliation}
\alias{d_defoliation}
\title{Example input data for r3PG simulations}
\format{
A named \code{list} with the following components:
\describe{
  \item{site}{A data frame with 1 row and 8 variables describing site conditions.}
  \item{species}{A data frame with rows corresponding to each species/cohort and variables such as initial biomass and modifiers.}
  \item{climate}{A data frame with monthly values of temperature, rainfall, radiation, frost days, etc.}
  \item{parameters}{A data frame containing species-specific physiological parameters.}
  \item{sizeDist}{A data frame describing the size distribution parameters for each species.}
  \item{thinning}{A data frame listing thinning interventions by species and age.}
  \item{defoliation}{A data frame listing defoliation events, including retained biomass fractions and recovery settings.}
}
}
\usage{
d_defoliation
}
\description{
A named list containing all required input tables for a 3PG simulation.
These include site conditions, species characteristics, climate, model parameters,
size distributions, thinning schedules, and defoliation events.
This dataset is particularly tailored to test defoliation
}
\examples{
data(d_defoliation)
str(d_defoliation$climate)

}
\seealso{
\code{\link{run_3PG}}, \code{\link{prepare_input}}
}
\keyword{datasets}
