% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_parameters}
\alias{d_parameters}
\title{Parameters input}
\format{
A \code{data frame} with 65 rows and x variables:
\describe{
  \item{parameter}{name of the parameter, must be consistent in naming with \code{\link{i_parameters}}}
  \item{Fagus sylvatica}{parameter values for species 1}
  \item{Pinus sylvestris}{parameter values for species 2}
}
}
\usage{
d_parameters
}
\description{
Table containing the information about parameters.
}
\keyword{datasets}
