% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_parameters_r}
\alias{d_parameters_r}
\title{Parameters input}
\format{
A \code{data frame} with 65 rows and x variables:
\describe{
  \item{parameter}{name of the parameter, must be consistent in naming with \code{\link{i_parameters}}}
  \item{Pinus sylvestris_1}{parameter values for species}
  \item{Pinus sylvestris_2}{parameter values for species}
  \item{Pinus sylvestris_3}{parameter values for species}
  \item{Pinus sylvestris_4}{parameter values for species}
  \item{Pinus sylvestris_5}{parameter values for species}
  \item{Pinus sylvestris_6}{parameter values for species}
  \item{Pinus sylvestris_7}{parameter values for species}
  \item{Pinus sylvestris_8}{parameter values for species}
  \item{Pinus sylvestris_9}{parameter values for species}
  \item{Pinus sylvestris_10}{parameter values for species}
  \item{Pinus sylvestris_11}{parameter values for species}
  \item{Pinus sylvestris_12}{parameter values for species}
  \item{Pinus sylvestris_13}{parameter values for species}
  \item{Pinus sylvestris_14}{parameter values for species}
  \item{Pinus sylvestris_15}{parameter values for species}
  \item{Pinus sylvestris_16}{parameter values for species}
  \item{Pinus sylvestris_17}{parameter values for species}
  \item{Pinus sylvestris_18}{parameter values for species}
  \item{Pinus sylvestris_19}{parameter values for species}
  \item{Pinus sylvestris_20}{parameter values for species}
  \item{Pinus sylvestris_21}{parameter values for species}
  \item{Pinus sylvestris_22}{parameter values for species}
  \item{Pinus sylvestris_23}{parameter values for species}
  \item{Pinus sylvestris_24}{parameter values for species}
  \item{Pinus sylvestris_25}{parameter values for species}
}
}
\usage{
d_parameters_r
}
\description{
Table containing the information about parameters.
}
\keyword{datasets}
