% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_parsQsoc}
\alias{d_parsQsoc}
\title{Parameters input}
\format{
A \code{data frame} with 65 rows and x variables:
\describe{
  \item{parameter}{name of the parameter, must be consistent in naming with \code{\link{i_parameters}}}
  \item{value}{parameter values}
}
}
\usage{
d_parsQsoc
}
\description{
Table containing the information about Q SOC parameters.
}
\keyword{datasets}
