% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_sizeDist}
\alias{d_sizeDist}
\title{sizeDist input}
\format{
A data frame with 47 rows and x variables:
\describe{
  \item{parameter}{name of the parameter, must be consistent in naming with \code{\link{i_sizeDist}}}
  \item{Fagus sylvatica}{parameter values for species 1}
  \item{Pinus sylvestris}{parameter values for species 2}
}
}
\usage{
d_sizeDist
}
\description{
Table containing the information about size distribution.
}
\keyword{datasets}
