% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_thinning_r}
\alias{d_thinning_r}
\title{Thinning input}
\format{
A \code{data frame} with 3 rows and 6 variables:
\describe{
  \item{species}{species or cohort id/name. It must be consistent with species names in \code{\link{d_species}}, \code{\link{d_parameters}} and \code{\link{d_sizeDist}} tables.}
  \item{age}{age when thinning is performed}
  \item{stems_n}{number of trees remaining after thinning}
  \item{stem}{type of thinning (above/below). Default is 1}
  \item{root}{type of thinning (above/below). Default is 1}
  \item{foliage}{type of thinning (above/below). Default is 1}
}
}
\usage{
d_thinning_r
}
\description{
Table containing the information about thinnings
}
\keyword{datasets}
