% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_climate}
\alias{data_climate}
\title{The weather data}
\format{A \code{three-dimentional array} with two \code{matrices} of dimentions 156 rows and 5 variables. The order of the collumns corresponds to the below listed items}
\usage{
data(data_climate)
}
\description{
A dataset containing the weather information.
}
\details{
\itemize{
  \item Tx: monthly mean daily maximum temperature
  \item Tn: monthly mean daily minimum temperature
  \item rain: monthly rainfall
  \item solarRad: monthly mean daily solar radiation
  \item frostDays: frost days per month
}
}
\keyword{datasets}
