% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_output_var}
\alias{data_output_var}
\title{Description of output variables}
\format{A vector of outpus. Please reffer to output ID (numerical value)}
\description{
Names and description of 3PGP output variables as described in \code{Table 4.} in \href{http://3pg.sites.olt.ubc.ca/files/2014/04/3PGpjs_UserManual.pdf}{original documentation}
}
\details{
\itemize{
  \item Stand attributes
   \itemize{
     \item 1: \code{StandAge} - Stand age (\eqn{yr})
     \item 2: \code{StemNo} - Stand stocking (\eqn{trees ha^-1})
     \item 3: \code{BasArea} - Stand basal area (\eqn{m^2 ha^-1})
     \item 4: \code{StandVol} - Stand volume excluding branch & bark (\eqn{m^3 ha^-1})
     \item 5: \code{avDBH} - Stand volume excluding branch & bark (\eqn{cm})
     \item 6: \code{MAI} - Mean annual volume increment (\eqn{m^3 ha^-1 yr^-1})
     }
  \item Canopy attributes
   \itemize{
    \item 7: \code{SLA} - Specific leaf area (\eqn{m^2 kg^-1})
    \item 8: \code{CanCover} - Fraction of ground area covered by canopy (\eqn{-})
    \item 9: \code{LAI} - Canopy LAI (\eqn{m^2 m^-2})
   }
  \item Biomass pools
   \itemize{
    \item 10: \code{WF} - Foliage biomass (\eqn{t_DM ha^-1})
    \item 11: \code{WR} - Root biomass (\eqn{t_DM ha^-1})
    \item 12: \code{WS} - Stem biomass, including branches and bark (\eqn{t_DM ha^-1})
    \item 13: \code{WL} - Accumulated litter fall (also TotalLitter) (\eqn{t_DM ha^-1})
    \item 14: \code{TotalW} - Total biomass (\eqn{t_DM ha^-1})
    \item 15: \code{AvStemMass} - Mean stem biomass per tree (\eqn{t_DM ha^-1})
    \item 16: \code{fracBB} - Fraction of stem biomass as branch and bark (\eqn{t_DM ha^-1})
   }
  \item Growth modifiers
   \itemize{
    \item 17: \code{fAge} - Age dependent modifier (\eqn{-})
    \item 18: \code{fVPD} - VPD dependent modifier (\eqn{-})
    \item 19: \code{fT} - Temperature dependent modifier (\eqn{-})
    \item 20: \code{fCalpha} - CO2 dependent modifier for quantum efficiency (\eqn{-})
    \item 21: \code{fCg} - CO2 dependent modifier canopy conductance (\eqn{-})
    \item 22: \code{fFrost} - Frost dependent modifier (\eqn{-})
    \item 23: \code{fSW} - Soil water dependent modifier (\eqn{-})
    \item 24: \code{fNutr} - Nutrition dependent modifier (\eqn{-})
    \item 25: \code{PhysMod} - Physiological modifier of canopy conductance (\eqn{-})
   }
  \item Biomass production and allocation
   \itemize{
    \item 26: \code{GPP} - Gross primary production in current period (\eqn{t_DM ha^-1})
    \item 27: \code{NPP} - Net primary production in current period (\eqn{t_DM ha^-1})
    \item 28: \code{RadInt} - Total solar radiation intercepted by canopy (\eqn{MJ m^-2 month^-1})
    \item 29: \code{alphaC} - Canopy quantum efficiency after modifiers (\eqn{mol mol^-1})
    \item 30: \code{epsilon} - Light utilisation efficiency based on total biomass (\eqn{g_DM MJ^-1})
    \item 31: \code{CVI} - Stem volume increment in current period (\eqn{m^3 ha^-1})
    \item 32: \code{m} - FR modifier of root biomass allocation (\eqn{-})
    \item 33: \code{pR} - Fraction of NPP allocated to roots (\eqn{-})
    \item 34: \code{pS} - Fraction of NPP allocated to stems (\eqn{-})
    \item 35: \code{pF} - Fraction of NPP allocated to foliage (\eqn{-})
    \item 36: \code{pFS} - Ratio of foliage to stem biomass allocation (\eqn{-})
    \item 37: \code{gammaF} - Current leaf litterfall rate (\eqn{month^-1})
   }
  \item Litterfall & root turnover
   \itemize{
    \item 38: \code{lossWF} - Litterfall turnover (\eqn{-})
    \item 39: \code{lossWR} - Toot turnover (\eqn{-})
   }
  \item Stem mortality
   \itemize{
    \item 40: \code{wSmax} - Max. mean tree stem mass at current stocking (\eqn{kg tree^-1})
    \item 41: \code{gammaN} - Density independent mortality rate (\eqn{month^-1})
    \item 42: \code{Mortality} - Number of stems dying in current period (\eqn{trees ha^-1})
   }
  \item Water use
   \itemize{
    \item 43: \code{supIrrig} - “Supplemental” irrigation to maintain Os >= Osm (\eqn{mm})
    \item 44: \code{RunOff} - Water loss due to run-off or drainage (\eqn{mm})
    \item 45: \code{fRainInt} - Fraction of rainfall intercepted by canopy (\eqn{-})
    \item 46: \code{RainInt} - Rainfall intercepted by canopy in current period (\eqn{mm})
    \item 47: \code{CanCond} - Canopy conductance (\eqn{m s^-1})
    \item 48: \code{WUE} - Water use efficiency (\eqn{gDM mm^-1})
    \item 49: \code{EvapTransp} - Evapotranspiration rate in current period (\eqn{mm})
    \item 50: \code{Transp} - Monthly transpiration rate in current period (\eqn{mm})
    \item 51: \code{ASW} - Available soil water (\eqn{mm})
   }
  \item Soil model
   \itemize{
    \item 52: \code{NEP} -  (\eqn{-})
    \item 53: \code{Rhet} -  (\eqn{-})
    \item 54: \code{Yr_C} - Refractory pool for carbon (\eqn{tC ha^−1})
    \item 55: \code{Yl_C} - Labile pool for carbon (\eqn{tC ha^−1})
    \item 56: \code{O_C} - Humified poll for carbon (\eqn{tC ha^−1})
    \item 57: \code{Yr_N} - Refractory pool for nitrogen (\eqn{tC ha^−1})
    \item 58: \code{Yl_N} - Labile pool for nitrogen (\eqn{tC ha^−1})
    \item 59: \code{O_N} - Humified poll for nitrogen (\eqn{tC ha^−1})
   }
}
}
\keyword{outputs}
