% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_param}
\alias{data_param}
\title{The parameter data}
\format{A \code{data frame} with 51 rows and 4 variables}
\usage{
data(data_param)
}
\description{
A dataset containing the parameters data, including medium, maximum and minimum.
}
\details{
\itemize{
  \item parName: parameter name
   \itemize{
     \item StemPower:
     \item aH: cnstant in the stem height relationship
     \item bW:
     \item klmax:
     \item alpha:
     \item Tmin: minimum temperature for growth
     \item krmax:
     \item gammaF1: maximum litterfall rate
     \item fN0: value of \code{fNutr} when \eqn{FR = 0}
     \item rg:
     \item rho1:
     \item gammaR: average monthly root turnover rate
     \item Topt: optimum temperature for growth
     \item MaxCond: maximum canopy conductance
     \item StemConst:
     \item pFS20: foliage:stem partitioning ratio D=20 cm
     \item pRn: minimum fraction of NPP to roots
     \item k: extinction coefficient for absorption of PAR by canopy
     \item fracBB1: branch and bark fraction for mature stands
     \item LAIgcx: LAI for maximum canopy conductance
     \item fullCanAge: age at canopy closure
     \item pRx: maximum fraction of NPP to roots
     \item CoeffCond: defines stomatal response to VPD
     \item pFS2: foliage:stem partitioning ratio D=2 cm
     \item hc:
     \item kF: number of days production lost for each frost day
     \item SLA1: specific leaf area for mature leaves
     \item tBB: age at which \eqn{fracBB = (fracBB0+fracBB1)/2}
     \item m0: value of \code{m} when \eqn{FR = 0}
     \item tSLA: age at which \eqn{SLA = (SLA0+SLA1)/2}
     \item Tmax: maximum temperature for growth
     \item MaxIntcptn: maximum proportion of rainfall evaporated from canopy
     \item fracBB0: branch and bark fraction at age 0
     \item SLA0: specific leaf area at age 0
     \item BLcond: canopy boundary layer conductance
     \item nAge: power of relative age in function for \code{fAge}
     \item tgammaF: age at which litterfall rate has median value
     \item MaxAge: maximum stand age used in age modifier
     \item rAge: relative age to give \eqn{fAge = 0.5}
     \item gammaF0: litterfall rate at \eqn{t = 0}
     \item komax:
     \item dmC:
     \item Yl_C_i:
     \item Yr_C_i:
     \item O_C_i:
     \item LAImaxIntcptn: LAI for maximum rainfall interception
     \item FR_esp:
     \item FR_ferc:
     \item FR_ferf:
     \item FR_feri:
     \item FR_ferif:
   }
  \item mode: medium value
  \item min: minimum value
  \item max: maximum value
}
}
\keyword{datasets}
