% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_param_alternative}
\alias{data_param_alternative}
\title{The parameter data ALTERNATIVE}
\format{A \code{data frame} with 51 rows and 4 variables}
\description{
The structure of parameters as in \href{http://3pg.sites.olt.ubc.ca/files/2014/04/3PGpjs_UserManual.pdf}{original documentation}
}
\details{
\itemize{
  \item Allometric relationships & partitioning
   \itemize{
     \item pFS2: foliage:stem partitioning ratio D=2 cm
     \item pFS20: foliage:stem partitioning ratio D=20 cm
     \item aWS: constant in the stem mass v. diam. relationship
     \item nWS: power in the stem mass v. diam. relationship
     \item pRx: maximum fraction of NPP to roots
     \item pRn: minimum fraction of NPP to roots
     }
  \item Litterfall & root turnover
   \itemize{
    \item gammaF1: maximum litterfall rate
    \item gammaF0: litterfall rate at t = 0
    \item tgammaF: age at which litterfall rate has median value
    \item gammaR: average monthly root turnover rate
    \item leafgrow: if deciduous, leaves are produced at end of this month
    \item leaffall: if deciduous, leaves all fall at start of this month
   }
}
}
\keyword{datasets}
