% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_site}
\alias{data_site}
\title{The site data}
\format{A \code{data frame} with 5 rows and 17 variables}
\usage{
data(data_site)
}
\description{
A dataset containing the site information about site.
}
\details{
\itemize{
  \item siteId: unique id of the single site. Can be character
  \item latitude: site latitude (-ve for the S hemisphere)
  \item nTrees: number of trees
  \item soilClass. soil class in accordance with Table 2 of \href{http://3pg.sites.olt.ubc.ca/files/2014/04/3PGpjs_UserManual.pdf}{model documentation}
  \item iAsw: ??
  \item minAsw: the minimum allowed available soil water (mm). If actual available soil water falls below a non-zero minimum, it is assumed the shortfall is made up by an external source
  \item maxAsw: maximum available water stored in the soil (mm)
  \item poolFraction:
  \item startAge: inicial age (\eqn{years}) of the site (Shall it be year?)
  \item endAge: the age (\eqn{years}) when tree is removed
  \item startMonth: inicial month of the site
  \item wfI: the individual stand foliage biomass pools (\eqn{t/ha}), i.e. at \code{startAge}, \code{startMonth}
  \item wrI: the individual stand root biomass pools (\eqn{t/ha}), i.e. at \code{startAge}, \code{startMonth}
  \item wsI: the individual stand stem (including branches and bark) biomass pools (\eqn{t/ha}), i.e. at \code{startAge}, \code{startMonth}
  \item nThinnings: identifies table of thinning events
  \item climId: an id of the climate data set \code{\link{data_climate}}
  \item fr:
}
}
\keyword{datasets}
