% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_thinning}
\alias{data_thinning}
\title{The thinning data}
\format{A \code{data frame} with 80 rows and 5 variables}
\usage{
data(data_thinning)
}
\description{
A dataset containing the thinning information about site.
}
\details{
\itemize{
  \item ageThinning: age (year and month as decimals) at which thinning is performed
  \item nTrees: number of trees remained
  \item f:
  \item r:
  \item s:
  \item siteId: the minimum allowed available soil water (mm). If actual available soil water falls below a non-zero minimum, it is assumed the shortfall is made up by an external source
}
}
\keyword{datasets}
