% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.config-data.R
\docType{data}
\name{example.config}
\alias{example.config}
\title{Example config}
\format{An list.}
\source{
\href{TODO}{TODO}
}
\usage{
data(example.config)
}
\description{
Data for setting up Ponderosa Pine at Kew OR isotope site
}
\examples{
data(example.config)
}
\references{
TODO...
}
\keyword{datasets}
