% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q_ss.R
\name{extend_time}
\alias{extend_time}
\title{This function repeats the climate time series for a specified length of time. It is used for initialization.}
\usage{
extend_time(climate, from, length)
}
\arguments{
\item{climate}{the original climatic time series.}

\item{from}{the start date of the desired new time series}

\item{length}{length in years of the desired new time series.}
}
\description{

}
\seealso{
\code{\link{C_init}}
}
