% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_3PG.R
\name{f_3PG}
\alias{f_3PG}
\title{The function to run the 3-PG in FORTRAN.}
\usage{
f_3PG(siteInputs, speciesInputs, forcingInputs, managementInputs,
  parameterInputs, sizeDistInputs, n_sp, n_m, n_man, t_t, settings = c(1L, 1L,
  1L, 0L, 0L), sp_names, df_out = TRUE)
}
\arguments{
\item{siteInputs}{\code{matrix} (dim=c(1,8)) containing the information about site level information}

\item{speciesInputs}{\code{matrix} (dim=c(n_sp,7)) containing the information about species level data. Each row corresponds to one species/layer.}

\item{forcingInputs}{\code{matrix} (dim=c(n_m,9)) containing the information about climatic forcing data (tmp_min, tmp_max, tmp_ave, prcp, srad, frost_days, vpd_day, co3, d13catm).}

\item{managementInputs}{\code{matrix} (dim=c(n_man,5,n_sp)) containing the information about management.}

\item{parameterInputs}{\code{matrix} (dim=c(82,n_sp)) parameters level data.}

\item{sizeDistInputs}{\code{matrix} (dim=c(30,n_sp)) bial level data}

\item{n_sp}{\code{integer} number of species}

\item{n_m}{\code{integer} number of simulated month}

\item{n_man}{\code{integer} number of maximum management interventions per species, dimention of managementInputs matrix (min=1)}

\item{t_t}{\code{vector} number of management interactions for each species}

\item{settings}{\code{vector} of integers  with all possible settings for the model.}

\item{sp_names}{\code{character vector} of species names}

\item{df_out}{\code{logical} if the output shall be in the 4-dimentional array (FALSE) or long data.frame (TRUE)}
}
\description{
This is the base implementation of the 3-PG model in Fortran. This function shall be used if you are sure that the input is in correct form. We always recomend to use \code{\link{run_3PG}}
}
\details{
Thre is no check for input quality and consistency. Therefore, it is highly recomended to check that your input is in the correct form
}
\examples{
f_site = as.matrix( site_eum )
f_species = as.matrix( species_eum[,-1] )
f_clim = as.matrix( climate_eum )
f_param = as.matrix( parameters_eum[,-1] )
f_sizeDist = as.matrix( sizeDist_eum[,-1] )

n_sp = nrow( f_species )
n_m = nrow( f_clim )
t_t = 0L
n_man = 1L
f_management = array(NA_real_, dim = c(1,5,n_sp))


out <- f_3PG(
  siteInputs = f_site,
  speciesInputs = f_species,
  forcingInputs = f_clim,
  managementInputs = f_management,
  parameterInputs = f_param,
  sizeDistInputs = f_sizeDist,
  n_sp = n_sp,
  n_m = n_m,
  n_man = n_man,
  t_t = t_t,
  settings = c(2L,2L,2L,0L,0L),
  sp_names = species_eum$species,
  df_out = FALSE)


out_long <- transf_out( out )
}
