% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_climate.R
\name{gen_climate}
\alias{gen_climate}
\title{The function generate climate table.}
\usage{
gen_climate(climate, n_years = 10, month_i = 1)
}
\arguments{
\item{climate}{a \code{data frame} or \code{matrix} containing the information about climatic forcing data. Shall have 12 rows, one for each month.
\itemize{
\item tmp_min: monthly mean daily minimum temperature (C).
\item tmp_max: monthly mean daily maximum temperature (C).
\item prcp: monthly rainfall (mm month-1).
\item srad: monthly mean daily solar radiation (MJ m-2 d-1).
\item frost_days: frost days per month (d month-1).
\item co2: required if calculate_d13c=1
\item d13catm: required if calculate_d13c=1
}}

\item{n_years}{for how many years to repeat the climate}

\item{month_i}{first month of the simulations.}
}
\description{
This function replicate the average climate for the required number of years.
}
\details{
This function shall be used in case user provide only average climate and want to repeate it for each year of simualtion.
}
\examples{
f_site = as.matrix( site_eum )
f_species = as.matrix( species_eum[,-1] )
f_clim = as.matrix( climate_eum )
f_param = as.matrix( parameters_eum[,-1] )
f_bias = as.matrix( bias_eum[,-1] )

n_sp = nrow( f_species )
n_m = nrow( f_clim )
t_t = 0L
n_man = 1L
f_management = array(NA_real_, dim = c(1,5,n_sp))


out <- f_3PG(
  siteInputs = f_site,
  speciesInputs = f_species,
  forcingInputs = f_clim,
  managementInputs = f_management,
  parameterInputs = f_param,
  biasInputs = f_bias,
  n_sp = n_sp,
  n_m = n_m,
  n_man = n_man,
  t_t = t_t,
  settings = c(2L,2L,2L,0L,0L),
  sp_names = species_eum$species,
  df_out = FALSE)


out_long <- transf_out( out )
}
