% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameters.R
\name{get_parameters}
\alias{get_parameters}
\title{Get parameter sets from published studies}
\usage{
get_parameters(mode = "overview", sp_names = NULL)
}
\arguments{
\item{mode}{A character string specifying the desired output. Must be one of:
\itemize{
  \item \code{"full"}: Full parameter set with all details.
  \item \code{"overview"}: Simplified table with key information.
  \item \code{"source"}: Summary of sources and references.
  \item \code{"comments"}: Parameter details along with notes and source comments.
  \item \code{"parameters"}: Parameters formatted for \code{\link{run_3PG}}.
  \item \code{"sizeDist"}: Size distribution parameters formatted for \code{\link{run_3PG}}.
}}

\item{sp_names}{A character vector of species names. Can include:
\itemize{
  \item Scientific names (e.g., \code{"Picea abies"}).
  \item Scientific names with a parameter set ID (e.g., \code{"Picea abies 37"}).
}
If \code{NULL}, retrieves all species (except for \code{mode = "source"}).}
}
\value{
A data frame containing parameter sets based on the selected mode and species.
}
\description{
Retrieves parameter sets for the 3-PG model from published studies, stored in the package's database.
}
\details{
This function accesses the parameter database stored in \code{\link{i_parameters_lit}}, which contains parameter sets from published studies.
\itemize{
  \item Use \code{mode = "overview"} to get a quick summary of available parameter sets.
  \item Use \code{mode = "parameters"} or \code{"sizeDist"} for formatted outputs ready for \code{\link{run_3PG}}.
  \item Use \code{sp_names} to filter results by species. Species names can optionally include the parameter set ID.
}
For detailed descriptions, consult the original publications referenced in \code{\link{i_parameters_lit}}.
}
\examples{
# see an overview of the existing parameter sets
get_parameters(mode = 'overview')
get_parameters(mode = 'overview', sp_names = c('Eucalyptus globulus', 'Pinus sylvestris' ) )

# see existing parameter sets and comments
get_parameters(mode = 'comments')
get_parameters(mode = 'comments', sp_names = c('Eucalyptus globulus', 'Pinus sylvestris' ) )

# see parameters and source information
get_parameters(mode = 'source')

# obtain parameter sets in for some species in the format required by run_3PG
get_parameters(mode = 'parameters', sp_names = c('Fagus sylvatica', 'Picea abies'))
get_parameters(mode = 'parameters', sp_names = 'Fagus sylvatica 34' )
get_parameters(mode = 'sizeDist', sp_names = c('Fagus sylvatica 9', 'Picea abies 37'))

# see parameter sets with full information (species, source, parameters)
get_parameters(mode = 'full')
get_parameters(mode = 'full', sp_names = c('Fagus sylvatica', 'Pinus radiata' ) )


}
\seealso{
\code{\link{run_3PG}}, \code{\link{i_parameters_lit}}
}
