% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StemMortality.R
\name{getMortality}
\alias{getMortality}
\title{Internal calculations for mortality}
\usage{
getMortality(oldN, oldW, mS, wSx1000, thinPower)
}
\description{
Determines the number of stems to remove to ensure the
  self-thinning rule is satisfied. It applies the Newton-Rhapson method
  to solve for N to an accuracy of 1 stem or less. To change this,
  change the value of "accuracy".#
}
