% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{i_parameters}
\alias{i_parameters}
\title{Information about parameters}
\format{
A data frame with 86 rows and 4 variables:
\describe{
  \item{parameter}{parameter name}
  \item{description}{description of the parameter}
  \item{unit}{unit}
  \item{default}{default value for E.globulus from original 3-PG}
}
}
\usage{
i_parameters
}
\description{
A dataset containing the parameters order and description.
}
\keyword{datasets}
