% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{i_parsQlitter}
\alias{i_parsQlitter}
\title{Information about Q litter parameters}
\format{
A data frame with 28 rows and 3 variables:
\describe{
  \item{parameter}{parameter name}
  \item{description}{description of the parameter}
  \item{unit}{unit}
  \item{default}{default value for E.globulus from original 3-PG}
}
}
\usage{
i_parsQlitter
}
\description{
A dataset containing the parameters order and description for Q litter.
}
\keyword{datasets}
