% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{i_parsQsoc}
\alias{i_parsQsoc}
\title{Information about Q SOC parameters}
\format{
A data frame with 6 rows and 3 variables:
\describe{
  \item{parameter}{parameter name}
  \item{description}{description of the parameter}
  \item{unit}{unit}
  \item{default}{default value for E.globulus from original 3-PG}
}
}
\usage{
i_parsQsoc
}
\description{
A dataset containing the parameters order and description for Q SOC.
}
\keyword{datasets}
