% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_config.R
\name{load_config}
\alias{load_config}
\title{Load 3PG configuration files}
\usage{
load_config(fpath_cfg)
}
\arguments{
\item{fpath_cfg}{String specifying the location of the configuration file.}
}
\value{
A list with 3PG configuration variables
}
\description{
This function enables loading of configuration files into R outside of the 
  primary run_3PG() function so that it can be modified and explored prior
  to running the model.
}
\examples{

}
\seealso{
\link[r3PG]{run_3PG}.
}
