% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CanopyProduction.R
\name{Pa_ppm}
\alias{Pa_ppm}
\title{Calculations for growth modifiers and productivity}
\usage{
Pa_ppm(x)
}
\description{
Internal functions for canopy productivity from light use and growth 
  modifiers for soil, vpd, etc.
}
