% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_parsSoc.R
\name{prepare_parsQsoc}
\alias{prepare_parsQsoc}
\title{Prepare Q parameters Litter table}
\usage{
prepare_parsQsoc(parameters = NULL)
}
\arguments{
\item{parameters}{table containing the information about parameters to be modified. Values that are not provided are replaced by defaults.
\itemize{
\item parameter: name of the parameter, must be consistent in naming with \code{\link{i_parsQsoc}}
\item species: each column must correspond to species/cohort id/name, as defined in \code{species} table
}}

\item{sp_names}{names of the species / cohorts used for the simulations.The `sp_names` must be identical to those from \code{species} table.}
}
\value{
a data.frame with 47 rows and columns corresponding to each species.
}
\description{
Prepares the soc parameters table used in Q, by either replicating the defaults or replicating defaults.
}
\details{
This function prepares the parameter table for \code{\link{run_3PG}}
}
\examples{
# replace some
prepare_parameters( parameters = d_parameters[1:4,],
  sp_names = c('Fagus sylvatica', 'Pinus sylvestris' ))

# Make default
prepare_parameters( parameters = NULL, sp_names = c('Quercus', 'Abies'))

}
\seealso{
\code{\link{run_3PG}}, \code{\link{prepare_input}}, \code{\link{prepare_sizeDist}}, \code{\link{prepare_thinning}}, \code{\link{prepare_climate}}
}
