% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qfun.R
\name{Q_dec}
\alias{Q_dec}
\title{Q decomposition function}
\usage{
Q_dec(inputs = inputs, fc, beta, e0, eta_11, u0, q0, delay, tmax, nSim)
}
\arguments{
\item{inputs}{the C inputs for the simulated cohorts}

\item{fc}{the C ratio of the input material}

\item{e0}{the efficiency of the microbial biomass}

\item{u0}{the growth rate of the microbial biomass}

\item{q0}{the initial quality of the C inputs coming to the cohort}

\item{delay}{the C inputs for the simulated cohorts}

\item{tmax}{the maximum diameter of the material particles or residuals}

\item{nSim}{time steps of the simulation}
}
\description{
A new implementation of the Q model included in this package.
This implementation is derived from Menichetti et al. (2021)
}
\details{
this function is just a demo version of what was then implemented in the FORTRAN code run by the
wrapper \code{\link{run_3PG}}, and its purpose is to simulate the decomposition of the outputs from the growth module 3PG.
}
\references{
Menichetti, L., Mäkinen, H., Stendahl, J., Ågren, G.I., Hyvönen, R., 2021. Modeling persistence of coarse woody debris residuals in boreal forests as an ecological property. Ecosphere 12. https://doi.org/10.1002/ecs2.3792
}
