% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q_ss.R
\name{Q_ss}
\alias{Q_ss}
\title{Steady state solution to the  Q decomposition function}
\usage{
Q_ss(fc, inputs, e0, eta11, b, u0, q0, Tmax)
}
\arguments{
\item{fc}{the C ratio of the input material}

\item{inputs}{the C inputs for the simulated cohorts}

\item{e0}{the efficiency of the microbial biomass}

\item{u0}{the growth rate of the microbial biomass}

\item{q0}{the initial quality of the C inputs coming to the cohort}

\item{tmax}{the maximum diameter of the material particles or residuals}
}
\description{

}
\details{
The function is used to initialize the decomposing pools in the C stocks simulation
}
\references{
Menichetti, L., Mäkinen, H., Stendahl, J., Ågren, G.I., Hyvönen, R., 2021. Modeling persistence of coarse woody debris residuals in boreal forests as an ecological property. Ecosphere 12. https://doi.org/10.1002/ecs2.3792
}
\seealso{
\code{\link{C_init}}
}
