% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r3pg_helpers}
\alias{r3pg_helpers}
\title{Help tables, need to be modified}
\format{1 \code{data frame}}
\usage{
data(r3pg_helpers)
}
\description{
Help tables, need to be modified
}
\keyword{datasets}
