% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q_ss.R
\name{sim_thinning}
\alias{sim_thinning}
\title{This function simulates a series of thinnings for Picea abies based on \code{\link{management_plan}}}
\usage{
sim_thinning(SI, lat, planted)
}
\arguments{
\item{SI}{The site index.}

\item{planted}{the date (year and month in format YY-M, as it is specified in \code{\link{run_3PG}})}

\item{Lat}{The latitude of the simulated site.}
}
\description{
This function simulates a series of thinnings for Picea abies based on \code{\link{management_plan}}
}
\seealso{
\code{\link{C_init}}, \code{\link{management_plan}}
}
