% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{site_eum}
\alias{site_eum}
\title{Site information}
\format{A data frame with 1 rows and 8 variables:
\describe{
  \item{latitude}{site latitude in the WGS84 coordinate system}
  \item{soil_class}{\code{integer} representing the soil class, according to table ...}
  \item{asw_i}{initial awailable soil water}
  \item{asw_max}{minimum awailable soil water}
  \item{asw_min}{maximum awailable soil water}
  \item{year_i}{\code{integer} year when the simulation will start}
  \item{month_i}{\code{integer} month when simulation will start, first month in the output file}
  \item{altitude}{altitude of the site, m a.s.l.}
}}
\usage{
site_eum
}
\description{
Information about site conditions for the example data.
}
\keyword{datasets}
