% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sizeDist_eum}
\alias{sizeDist_eum}
\title{sizeDist information}
\format{A data frame with 47 rows and x variables:
\describe{
  \item{parameter}{name of the parameter}
  \item{Fagus sylvatica}{parameter values for species 1}
  \item{Pinus sylvestris}{parameter values for species 2}
}}
\usage{
sizeDist_eum
}
\description{
Containing the information about bias parameters.
}
\keyword{datasets}
