% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sizeDist_info}
\alias{sizeDist_info}
\title{sizeDist table information}
\format{A data frame with 30 rows and 3 variables:
\describe{
  \item{parameter}{parameter name}
  \item{description}{description of the parameter}
  \item{unit}{unit}
  \item{default}{default value equal to 0}
}}
\usage{
sizeDist_info
}
\description{
A dataset containing the parameters order and description.
}
\keyword{datasets}
