% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_climate.R
\name{sub_climate}
\alias{sub_climate}
\title{The function to subset climate table.}
\usage{
sub_climate(climate, from = "2000-04", to = "2010-11")
}
\arguments{
\item{climate}{a \code{data frame} or \code{matrix} containing the information about climatic forcing data. In case you want to replicate average climate it shall have exactly 12, one for each month. In case one want to subset, it shall contain the period for subsetting and two additional columns: year and month.  Additionally can include: tmp_ave, c02, d13catm. The minimum required columns are listed below.
\itemize{
\item year: numeric, year of observation (only reguired for subsetting.
\item month: numeric, months of observation (only reguired for subsetting.
\item tmp_min: monthly mean daily minimum temperature (C).
\item tmp_max: monthly mean daily maximum temperature (C).
\item prcp: monthly rainfall (mm month-1).
\item srad: monthly mean daily solar radiation (MJ m-2 d-1).
\item frost_days: frost days per month (d month-1).
}}

\item{from}{from which date climate data shall be included. Shall be provided as character, in form of year-month. E.g. "2000-01"}

\item{to}{to which date climate data shall be included. Shall be provided as character, in form of year-month. E.g. "2009-12", will include December 2009 as last simulation month}
}
\description{
This function prepare the climate table, by either replicating the average climate for the required number of years or subsetting.
}
\details{
This function shall be used in case user provide only average climate and want to repeate it for each year of simualtion.
}
\examples{
climate = matrix(rnorm(60), ncol = 5)
colnames(climate) = c("tmp_min", "tmp_max", "prcp", "srad", "frost_days")

sub_climate( climate, from = '2000-04', to = '2010-11')
}
