% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{thinn_eum}
\alias{thinn_eum}
\title{Management information}
\format{A data frame with 3 rows and 6 variables:
\describe{
  \item{species}{\code{integer} species id}
  \item{age}{age at which management is done}
  \item{n_trees}{number of trees remaining after management}
  \item{foliage}{type of thinning (above/below). Default is 1}
  \item{root}{type of thinning (above/below). Default is 1}
  \item{stem}{type of thinning (above/below). Default is 1}
}}
\usage{
thinn_eum
}
\description{
Information about management
}
\keyword{datasets}
