% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q_ss.R
\name{u0_calc}
\alias{u0_calc}
\title{Scaling of the microbial activity term}
\usage{
u0_calc(latitude)
}
\arguments{
\item{lat}{latitude, self explanatory}
}
\description{

}
\details{
The function is used to rescale the microbial activity (and therefore the kinetic of decomposition)
based on latitude
}
\references{
Menichetti, L., Mäkinen, H., Stendahl, J., Ågren, G.I., Hyvönen, R., 2021. Modeling persistence of coarse woody debris residuals in boreal forests as an ecological property. Ecosphere 12. https://doi.org/10.1002/ecs2.3792
}
\seealso{
\code{\link{Q_ss}}
}
