% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climate_eum}
\alias{climate_eum}
\title{Climate forcing data}
\format{A data frame with 156 rows and 7 variables:
\describe{
  \item{Tmin}{minimum monthly temperature}
  \item{Tmax}{maximum monthly temperature}
  \item{Rain}{total monthly precipittation}
  \item{sRad}{solar radiation}
  \item{fDays}{number of frost days}
  \item{co2}{atmospheric co2}
  \item{d13catm}{d13catm}
}}
\usage{
climate_eum
}
\description{
Monthly values for climatic data
}
\keyword{datasets}
