% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_param}
\alias{data_param}
\title{The parameter data}
\format{A \code{data frame} with 51 rows and 4 variables}
\usage{
data(data_param)
}
\description{
The order of the parameters corresponds to the \code{Table 4} in \href{http://3pg.sites.olt.ubc.ca/files/2014/04/3PGpjs_UserManual.pdf}{original documentation}, but in the \code{data frame} to numbers in front of the parameter
}
\details{
\itemize{
  \item parName: parameter name
\itemize{
  \item Allometric relationships & partitioning
   \itemize{
     \item 24: \code{pFS2} - Foliage:stem partitioning ratio \eqn{D = 2} cm
     \item 16: \code{pFS20} - Foliage:stem partitioning ratio \eqn{D = 20} cm
     \item -: \code{aWS} - Constant in the stem mass v. diam. relationship (alt. stemConst)
     \item -: \code{nWS} - Power in the stem mass v. diam. relationship (alt. stemPower)
     \item 22: \code{pRx} - Maximum fraction of NPP to roots
     \item 17: \code{pRn} - Minimum fraction of NPP to roots
     }
  \item Litterfall & root turnover
   \itemize{
    \item 8: \code{gammaF1} - Maximum litterfall rate
    \item 40: \code{gammaF0} - Litterfall rate at \eqn{t = 0}
    \item 37: \code{tgammaF} - Age at which litterfall rate has median value
    \item 12: \code{gammaR} - Average monthly root turnover rate
    \item -: \code{leafgrow} - If deciduous, leaves are produced at end of this month
    \item -: \code{leaffall} - If deciduous, leaves all fall at start of this month
   }
  \item Temperature modifier \code{fT}
   \itemize{
    \item 6: \code{Tmin} - Minimum temperature for growth
    \item 13: \code{Topt} - Optimum temperature for growth
    \item 31: \code{Tmax} - Maximum temperature for growth
   }
  \item Frost modifier \code{fFrost}
   \itemize{
    \item 26: \code{kF} - Number of days production lost for each frost day
   }
  \item Fertility modifiers
   \itemize{
    \item 29: \code{m0} - Value of \code{m} when \eqn{FR = 0}
    \item 9: \code{fN0} - Value of \code{fNutr} when \eqn{FR = 0}
    \item -: \code{fNn} - Power of \eqn{(1-FR)} in \code{fN}
   }
  \item Age modifier
   \itemize{
    \item 38: \code{MaxAge} - Maximum stand age used in age modifier
    \item 36: \code{nAge} - Power of relative age in function for \code{fAge}
    \item 39: \code{rAge} - Relative age to give \eqn{fAge = 0.5}
   }
  \item Stem mortality and self-thinning
   \itemize{
    \item -: \code{gammaN0} -
   }
  \item Conductance
   \itemize{
    \item -: \code{MinCond} -
    \item 14: \code{MaxCond} - Maximum canopy conductance
    \item 20: \code{LAIgcx} - LAI for maximum canopy conductance
    \item 23: \code{CoeffCond} - Defines stomatal response to VPD
    \item 35: \code{BLcond} - Canopy boundary layer conductance
   }
  \item Specific leaf area
   \itemize{
    \item 34: \code{SLA0} - Specific leaf area at age 0
    \item 27: \code{SLA1} - Specific leaf area for mature leaves
    \item 30: \code{tSLA} - Age at which \eqn{SLA = (SLA0+SLA1)/2}
   }
  \item Rainfall interception
   \itemize{
    \item 32: \code{MaxIntcptn} - Maximum proportion of rainfall evaporated from canopy
    \item 46: \code{LAImaxIntcptn} - LAI for maximum rainfall interception
   }
  \item Light interception, production and respiration
   \itemize{
    \item 18: \code{k} - Extinction coefficient for absorption of PAR by canopy
    \item 21: \code{fullCanAge} - Age at canopy closure
    \item 5: \code{alpha} - Maximum canopy quantum efficiency
    \item -: \code{Y} -
   }
  \item Branch and bark fraction \code{(fracBB)}
   \itemize{
    \item 33: \code{fracBB0} - Branch and bark fraction at age 0
    \item 19: \code{fracBB1} - Branch and bark fraction for mature stands
    \item 28: \code{tBB} - Age at which \eqn{fracBB = (fracBB0+fracBB1)/2}
   }
  \item Stem height allometric relationship
   \itemize{
    \item 2: \code{aH} - Constant in the stem height relationship
   }
  \item NOT DOCUMENTED
   \itemize{
    \item 1: \code{StemPower} -
    \item 3: \code{bW} -
    \item 4: \code{klmax} -
    \item 7: \code{krmax} -
    \item 10: \code{rg} -
    \item 11: \code{rho1} -
    \item 15: \code{StemConst} -
    \item 25: \code{hc} -
    \item 41: \code{komax} -
    \item 42: \code{dmC} -
    \item 43: \code{Yl_C_i} -
    \item 44: \code{Yr_C_i} -
    \item 45: \code{O_C_i} -
    \item 47: \code{FR_esp} -
    \item 48: \code{FR_ferc} -
    \item 49: \code{FR_ferf} -
    \item 50: \code{FR_feri} -
    \item 51: \code{FR_ferif} -
   }
   }
  \item mode: medium value
  \item min: minimum value
  \item max: maximum value
}
}
\keyword{datasets}
