% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.clim-data.R
\docType{data}
\name{example.clim}
\alias{example.clim}
\title{Example climate}
\format{An list.}
\source{
\href{TODO}{TODO}
}
\usage{
data(example.clim)
}
\description{
Cliamte data for Kew OR isotope site
}
\examples{
data(example.clim)
}
\references{
TODO...
}
\keyword{datasets}
