% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameters.R
\name{get_parameters}
\alias{get_parameters}
\title{Get parameter sets}
\usage{
get_parameters(df_out = "overview", sp_names = NULL)
}
\arguments{
\item{df_out}{must be one of the following: overview, source, full, run.}

\item{sp_names}{names of the species. The `sp_names` must be either a scientific name
(Picea abies) or a scientific name with a integer (Picea abies 37) with df_out = run.
If unsure, consider using first the option df_out = overview with no species.
See details below for more information.}
}
\value{
a data.frame with parameter sets for all available species, or only the
 requested species, if sp_names is not null.
}
\description{
Gets parameters sets for species from published studies
}
\details{
This function access the internal parameter database stored in the package. If 
df_out = overview, a simplified table with all existing parameter sets will be returned. 
A extended version including information related to the source can be obtained
with df_out = source. The complete version of this table (species,source, comments,
parameters) will be returned if df_out = full. 
Alternatively, with df_out = run, it is possible to
obtain only the parameters in the format required by \code{\link{run_3PG}}.
Passing species names with sp_names means selecting from the table
the desired species. Please note that one species might have more than one 
available parameter sets.

The parameter sets were obtained from published studies. Basic information about the
sets and the corresponding studies is provided in the table. For further information, 
please consider checking the original publication.
}
\examples{
# see an overview of the existing parameter sets
get_parameters(df_out = 'overview')
get_parameters(df_out = 'overview', sp_names = c('Eucalyptus globulus', 'Pinus sylvestris' ) )

# see parameters and source information
get_parameters(df_out = 'source')
get_parameters(df_out = 'source', sp_names = c('Fagus sylvatica', 'Salix dasyclados' ) )

# see parameter sets with full information (species, source, parameters)
get_parameters(df_out = 'full')
get_parameters(df_out = 'full', sp_names = c('Fagus sylvatica', 'Pinus radiata' ) )

# obtain parameter sets in for some species in the format required by run_3PG
get_parameters(df_out = 'run', sp_names = c('Fagus sylvatica', 'Picea abies'))
get_parameters(df_out = 'run', sp_names = 'Fagus sylvatica 34' )
get_parameters(df_out = 'run', sp_names = c('Fagus sylvatica 9', 'Picea abies 37'))


}
\seealso{
\code{\link{run_3PG}}, \code{\link{i_parameters}}
}
