% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{i_output}
\alias{i_output}
\title{Information about model outputs}
\format{
A data frame with 150 rows and 7 variables:
\describe{
  \item{group_id}{serial number of the group}
  \item{variable_id}{serial number of the variable}
  \item{variable_group}{group name to which variable belongs}
  \item{variable_name}{variable name as named in output}
  \item{description}{description of the variable}
  \item{unit}{unit of the variable}
  \item{variable_vba}{corresponding name of the variable as output from Excel version of 3-PGmix}
}
}
\usage{
i_output
}
\description{
A dataset containing the list of output variables and their description.
}
\keyword{datasets}
