% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{i_parameters_lit}
\alias{i_parameters_lit}
\title{Information about literature parameters}
\format{
A data frame with XX:
\describe{
  \item{parameter}{parameter name}
}
}
\usage{
i_parameters_lit
}
\description{
A dataset containing the parameters order and description.
}
\keyword{datasets}
