% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q_ss.R
\name{management_plan}
\alias{management_plan}
\title{The function for determining the appropriate management plan.}
\usage{
management_plan(SI, Lat)
}
\arguments{
\item{SI}{The site index.}

\item{Lat}{The latitude of the simulated site.}
}
\value{
a data frame with :
}
\description{
The management.plan functions contains a series of vectors with management data from Sweden.
It selects the most appropriate values based on site index and latitude, and determines the
values for the management plan.
}
\seealso{
\code{\link{C_init}}
}
