% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parameters_eum}
\alias{parameters_eum}
\title{Parameters information}
\format{A data frame with 65 rows and x variables:
\describe{
  \item{parameter}{name of the parameter}
  \item{Fagus sylvatica}{parameter values for species 1}
  \item{Pinus sylvestris}{parameter values for species 2}
}}
\usage{
parameters_eum
}
\description{
Containing the information about parameters.
}
\keyword{datasets}
