% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_climate.R
\name{prepare_climate}
\alias{prepare_climate}
\title{Subsets or replicate a climate data}
\usage{
prepare_climate(climate, from = "2000-04", to = "2010-11")
}
\arguments{
\item{climate}{table containing the information about monthly values for climatic data. If the climate table have exactly 12 rows it will be replicated for the number of years and months specified by \code{from} - \code{to}. Otherwise, it will be subsetted to the selected time period. It is required to include columsn \code{year} and \code{month} in case climate table shall be subsetted to selected period. The minimum required columns are listed below, but additionally you can include: tmp_ave, c02, d13catm. Please reffer to \code{\link{d_climate}} for example.
\itemize{
\item year: year of observation (only reguired for subsetting) (numeric).
\item month: months of observation (only reguired for subsetting) (numeric).
\item tmp_min: monthly mean daily minimum temperature (C).
\item tmp_max: monthly mean daily maximum temperature (C).
\item tmp_ave: monthly mean daily mean temperature (C) (optional).
\item prcp: monthly rainfall (mm month-1).
\item srad: monthly mean daily solar radiation (MJ m-2 d-1).
\item frost_days: frost days per month (d month-1).
\item co2: required if calculate_d13c=1 (optional).
\item d13catm: required if calculate_d13c=1 (optional).
}}

\item{from}{year and month indicating the start of simulation. Provided in form of year-month. E.g. "2000-01".}

\item{to}{year and month indicating the end of simulation. Provided in form of year-month. E.g. "2009-12", will include December 2009 as last simulation month.}
}
\description{
This function prepares the climate table, by either replicating the average climate for the required number of years, or by subsetting from the longer time-series of climate data.
}
\details{
This function prepares the climate table for the model.

In case a user provide only average climate, this is replicated for the desired simulation period.

In case a larger climate file is provided, the simulation period is selected from this.
}
\examples{
# subsetting climate data
prepare_climate( climate = d_climate, from = '2003-04', to = '2010-11')

# replicating climate data
climate = matrix(rnorm(60), ncol = 5)
colnames(climate) = c("tmp_min", "tmp_max", "prcp", "srad", "frost_days")

prepare_climate( climate = climate, from = '2000-04', to = '2010-11')
}
