% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_defoliation.R
\name{prepare_defoliation}
\alias{prepare_defoliation}
\title{Check and prepare defoliation information}
\usage{
prepare_defoliation(defoliation = NULL, sp_names = c("Fagus sylvatica",
  "Pinus sylvestris"))
}
\arguments{
\item{defoliation}{A data frame containing defoliation information. If no defoliation is required, set to \code{NULL}. The following columns are required:
\itemize{
  \item \code{species}: Species or cohort ID/name.
  \item \code{age}: Age (years) at which defoliation occurs (numeric).
  \item \code{def_type}: Defoliation type: 1 - prune; 2 - copice; 3 - epicormic; 4 - stand replacing
  \item \code{stem_retained}: Proportion of stem mass retained after defoliation (0 to 1).
  \item \code{foliage_retained}: Proportion of foliage mass retained after defoliation (0 to 1).
  \item \code{root_retained}: Proportion of root mass retained after defoliation (0 to 1).
  \item \code{stem}: Fraction of average tree stem mass of killed trees (numeric, default 1).
  \item \code{def_recover_t}: Time (months) to recover from defoliation (numeric). After this time growth is only from npp (not non-structural carbohydrates), and biomass partitioning returns to normal.
  \item \code{prop_carbs}: Proportion of pre-defoliation non-structural carbohydrates used to regenerate foliage (0 to 1).
  \item \code{prop_npp}: Proportion of new photosynthate allocated to foliage (0 to 1).
}}

\item{sp_names}{A character vector of species or cohort names used in the simulation. Must match \code{species} names in the input.}
}
\value{
A 3-dimensional array of defoliation parameters, indexed by [event, parameter, species].
}
\description{
Prepares the defoliation table for simulations, ensuring data consistency and completeness.
}
