% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_input.R
\name{prepare_input}
\alias{prepare_input}
\title{Check and prepare input for running 3-PG model}
\usage{
prepare_input(site, species, climate, thinning = NULL, parameters = NULL,
  size_dist = NULL, settings = NULL)
}
\arguments{
\item{site}{table containing the information about site conditions.
\itemize{
\item latitude: site latitude in the WGS84 coordinate system.
\item altitude: site altitude, m a.s.l.
\item soil_class: soil class, according to table 2 user manual of 3PGpjs. 1 - Sandy; 2 - Sandy loam; 3 - Clay loam; 4 - Clay; 0 - No effect of asw on production.
\item asw_i: initial awailable soil water (mm).
\item asw_min: minimum awailable soil water (mm).
\item asw_max: maximum awailable soil water (mm).
\item from: year and month indicating the start of simulation. Provided in form of year-month. E.g. "2000-01".
\item to: year and month indicating the end of simulation. Provided in form of year-month. E.g. "2009-12", will include December 2009 as last simulation month
}}

\item{species}{table containing the information about species level data. Each row corresponds to one species/cohort.
\itemize{
\item species: species or cohort id/name. It shall be consistent with species names in \code{thinning}, \code{parameters} and \code{sizeDist} tables.
\item planted: year and month indicating when species was planted. Provided in form of year-month. E.g. "2000-01".
\item fertility: soil fertility for a given species. Range from 0 to 1.
\item stems_n: number of trees per ha.
\item biom_stem: stem biomass for a given species (T/ha).
\item biom_root: root biomass for a given species (T/ha).
\item biom_foliage: initial foliage biomass (T/ha). If this is a leafless period provide a foliage biomass in summer.
}}

\item{climate}{table containing the information about monthly values for climatic data. If the climate table have exactly 12 rows it will be replicated for the number of years and months specified by \code{from} - \code{to}. Otherwise, it will be subsetted to the selected time period. More details about preparin climate data are at \code{\link{prepare_climate}}.
\itemize{
\item year: year of observation (only reguired for subsetting) (optional).
\item month: months of observation (only reguired for subsetting) (optional).
\item tmp_min: monthly mean daily minimum temperature (C).
\item tmp_max: monthly mean daily maximum temperature (C).
\item tmp_ave: monthly mean daily maximum temperature (C) (optional).
\item prcp: monthly rainfall (mm month-1).
\item srad: monthly mean daily solar radiation (MJ m-2 d-1).
\item frost_days: frost days per month (d month-1).
\item vpd_day: frost days per month (mbar) (optional).
\item co2: required if calculate_d13c=1 (optional)
\item d13catm: required if calculate_d13c=1 (optional)
}}

\item{thinning}{table containing the information about thinnings. In case there is no management it shall be equall to \code{NULL}.
\itemize{
\item species: species or cohort id/name. It shall be consistent with species names in \code{species}, \code{parameters} and \code{sizeDist} tables.
\item age: age when thinning is performed.
\item stems_n: number of trees remaining after management
\item foliage: type of management (above/below). Default is 1.
\item root: type of management (above/below). Default is 1.
\item stem: type of management (above/below). Default is 1.
}}

\item{parameters}{table containing the information about parameters to be modified. Values that are not provided are replaced by defaults.
\itemize{
\item parameter: name of the parameter, shall be consistent in naming with \code{\link{i_parameters}}
\item species: each column shall correspond to species/cohort id/name, as defined in \code{species} table
}}

\item{size_dist}{table containing the information about size distribution to be modified. Values that are not provided are replaced by defaults.
\itemize{
\item parameter: name of the parameter, shall be consistent in naming with \code{\link{i_sizeDist}}
\item species: each column shall correspond to species/cohort id/name, as defined in \code{species} table
}}

\item{settings}{a list with settings for the model. Values that are not provided are replaced by defaults.
\itemize{
\item light_model: `1` - 3-PGpjs (default); `2` - 3-PGmix
\item transp_model: `1` - 3-PGpjs (default); `2` - 3-PGmix
\item phys_model:  `1` - 3-PGpjs (default); `2` - 3-PGmix
\item height_model: `1` - linear (default); `2` - non-linear
\item correct_bias: `0` - no (default); `1` - yes
\item calculate_d13c: `0` - no (default); `1` - yes
}}
}
\description{
This function check and prepare all input tables to be used in \code{\link{run_3PG}}.
}
\details{
This function checks and prepares the input data for the \code{\link{run_3PG}}. The output is a list with 7 tables. Each of them corresponds to the one from input.
}
\examples{
prepare_input( site = d_site, species = d_species, climate = d_climate, d_thinning)
}
