% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_parameters.R
\name{prepare_parameters}
\alias{prepare_parameters}
\title{Prepare parameters table}
\usage{
prepare_parameters(parameters = NULL, sp_names = c("Fagus sylvatica",
  "Pinus sylvestris"))
}
\arguments{
\item{parameters}{table containing the information about parameters to be modified. Values that are not provided are replaced by defaults.
\itemize{
\item parameter: name of the parameter, shall be consistent in naming with \code{\link{i_parameters}}
\item species: each column shall correspond to species/cohort id/name, as defined in \code{species} table
}}

\item{sp_names}{names of the species / cohorsts used for the simulations.The `sp_names` shall be identical to those from \code{species} table.}
}
\description{
This function prepares the parameters table, by either replicating the defaults or replicating defaults for each of the species.
}
\details{
This function prepares the parameter table for the model.
}
\examples{
# replace some
prepare_parameters( parameters = d_parameters[1:4,],
  sp_names = c('Fagus sylvatica', 'Pinus sylvestris' ))

# Make default
prepare_parameters( parameters = NULL, sp_names = c('Quercus', 'Abies'))

}
