% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_site.R
\name{prepare_site}
\alias{prepare_site}
\title{Check the site data for consistency}
\usage{
prepare_site(site)
}
\arguments{
\item{site}{A data frame containing site-level data. It must contain exactly one row with the following columns:
\itemize{
  \item \code{latitude}: Site latitude in the WGS84 coordinate system (degrees, range: [-90, 90]).
  \item \code{elevation}: Site elevation (meters above sea level, range: [0, 4000]).
  \item \code{soil_class}: Soil class as per 3PG User Manual Table 27:
    \itemize{
      \item 1: Clay
      \item 2: Clay loam
      \item 3: Loam
      \item 4: Loamy sand
      \item 5: Sand
      \item 6: Sandy clay
      \item 7: Sandy clay loam
      \item 8: Sandy loam
      \item 9: Silt
      \item 10: Silty clay
      \item 11: Silty clay loam
      \item 12: Silty loam
      \item 0: Uses cθ and nθ provided in the “parameters” input
    }
  \item \code{asw_i}: Initial available soil water (mm, must be >= 0).
  \item \code{asw_min}: Minimum available soil water (mm, must be >= 0).
  \item \code{asw_max}: Maximum available soil water (mm, must be >= 0).
  \item \code{from}: Start of simulation period (year-month, e.g., "2000-01").
  \item \code{to}: End of simulation period (year-month, e.g., "2009-12"). The simulation includes the entire month of December 2009.
  \item \code{lt_mod_mths}: Number of months considered to be long-term when calculating the long-term average modifier values for density-dependent mortality.
  \item \code{st_Power}: Power in self-thinning rule when mort_model = 2
  \item \code{st_Intercept}: Intercept for self-thinning rule (mort_model = 2)
  \item \code{st_fN}: Power in fertility modifier when (mort_model = 2)
  \item \code{st_fT}: Power in temperature modifier when (mort_model = 2)
  \item \code{st_fPhys}: Power in physmod modifier when (mort_model = 2)
  \item \code{beta0}: Constant for self-thinning when mort_model = 3
  \item \code{betaB}: Power for B when mort_model = 3
  \item \code{betaN}: Power in tree density when mort_model = 3
  \item \code{betafN}: Power in nutrition modifier when mort_model = 3
  \item \code{betafT}: Power in temperature modifier when mort_model = 3
  \item \code{betafPhys}: Power in physmod modifier when mort_model = 3


}}
}
\value{
A validated and prepared data frame with one row.
}
\description{
Validates and prepares the site table for use in the 3PG model. Ensures that the input information is consistent and formatted correctly.
}
\details{
This function validates the site table for \code{\link{run_3PG}} and ensures that all required fields are consistent.
}
\examples{
# Example: Valid site data
# r3PG default data
prepare_site(site = d_mixture$site)

# example data
d_site <- data.frame(
  latitude = 46.8,
  elevation = 1200,
  soil_class = 2,
  asw_i = 150,
  asw_min = 100,
  asw_max = 200,
  from = "2000-01",
  to = "2009-12"
)

# Validate and prepare site data
prepared_site <- prepare_site(d_site)

# View the prepared site data
print(prepared_site)
}
\seealso{
\code{\link{run_3PG}}, \code{\link{prepare_input}}, \code{\link{prepare_parameters}}, \code{\link{prepare_sizeDist}}, \code{\link{prepare_thinning}}
}
