% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_sizeDist.R
\name{prepare_sizeDist}
\alias{prepare_sizeDist}
\title{Prepare parameters table}
\usage{
prepare_sizeDist(size_dist = NULL, sp_names = c("Fagus sylvatica",
  "Pinus sylvestris"))
}
\arguments{
\item{size_dist}{table containing the information about size distribution to be modified. Values that are not provided are replaced by defaults.
\itemize{
\item parameter: name of the parameter, shall be consistent in naming with \code{\link{i_sizeDist}}.
\item species: each column shall correspond to species/cohort id/name, as defined in \code{species} table.
}}

\item{sp_names}{names of the species / cohorsts used for the simulations.The `sp_names` shall be identical to those from \code{species} table.}
}
\description{
This function prepares the parameters table, by either replicating the defaults or replicating defaults for each of the species.
}
\details{
This function prepares the parameter table for the model.
}
\examples{
# replace some
prepare_sizeDist( size_dist = d_sizeDist[1:4,],
  sp_names = c('Fagus sylvatica', 'Pinus sylvestris' ))

# Make default
prepare_sizeDist( size_dist = NULL, sp_names = c('Quercus', 'Abies'))

}
