% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_thinning.R
\name{prepare_thinning}
\alias{prepare_thinning}
\title{Check and prepare management information.}
\usage{
prepare_thinning(thinning = NULL, sp_names = c("Fagus sylvatica",
  "Pinus sylvestris"))
}
\arguments{
\item{thinning}{a \code{data frame} or \code{matrix} containing the information about management. In case there is no management it shall be equall to \code{NULL}. The following columns are required:
\itemize{
\item species: species or cohort id/name.
\item age: age at which management is done.
\item n_trees: number of trees remaining after management
\item foliage: type of management (above/below). Default is 1.
\item root: type of management (above/below). Default is 1.
\item stem: type of management (above/below). Default is 1.
}}

\item{sp_names}{names of the species / cohorsts used for the simulations. This is required to account if `thinning=NULL` or if not all species are indicated in the `thinning` table. The `sp_names` shall be identical to those from \code{species} table.}
}
\description{
This function prepares the management table and ckeck for consistensy.
}
\details{
This function prepares the thinning table for the model.

In case there is no management it will return empty 3-d array.

In case there will be management it will return 3-d array, where one dimention correspond to each species.
}
\examples{
prepare_thinning( thinning = NULL, sp_names = c('Quercus', 'Abies'))

prepare_thinning( thinning = thinn_eum, sp_names = c('Fagus sylvatica', 'Pinus sylvestris'))
}
