% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r3PG.R
\docType{package}
\name{r3PG}
\alias{r3PG}
\title{r3PG -}
\description{
add description here
}
\examples{
out <- run_3PG(
  siteInputs = site_eum,
  speciesInputs = species_eum,
  forcingInputs = climate_eum,
  managementInputs = NULL,
  parameterInputs = parameters_eum,
  biasInputs = bias_eum,
  settings = list(light_model = 2, transp_model = 2, phys_model = 2,
                  correct_bias = 0, calculate_d13c = 0))


out_long <- transf_out( out )
}
\seealso{
\code{\link{run_3PG}}
}
