% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r3PG.R
\docType{package}
\name{r3PG}
\alias{r3PG}
\title{r3PG – an R package for forest growth simulation using 3-PG process-based model}
\description{
The r3PG serves as a flexible and easy-to-use interface for the 3-PGpjs and the 3-PGmix model written in Fortran.The package, allows for fast and easy interaction with the model, and Fortran re-implementation facilitates computationally intensive sensitivity analysis and calibration. The user can flexibly switch between various options and submodules, to use the original 3-PGpjs model version for monospecific, even-aged and evergreen forests and the 3-PGmix model, which can also simulate multi-cohort stands (e.g. mixtures, uneven-aged) that contain deciduous species.
}
\examples{
out <- run_3PG(
  site = d_site,
  species = d_species,
  climate = d_climate,
  thinning = d_thinning,
  parameters = d_parameters,
  size_dist = d_sizeDist,
  settings = list(light_model = 2, transp_model = 2, phys_model = 2,
                  correct_sizeDist = 1, calculate_d13c = 0),
  check_input = TRUE, df_out = TRUE) # note that default is TRUE

str(out) # List output format
}
\seealso{
\code{\link{run_3PG}}
}
