% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3PG.R
\name{r3pgn}
\alias{r3pgn}
\title{Define the function to run 3PGN in R}
\usage{
r3pgn(siteData, weather, thinning = NA, parameters, outputs = NA)
}
\arguments{
\item{nvariables}{needs to be documented later}
}
\description{
Define the function to run 3PGN in R
}
\details{
This is the model
}
\examples{
firstRun <- r3pgn(siteData = siteData,weather=weather,parameters = pars[,2])
print(firstRun)
plot(firstRun)

plot(firstRun$output[,1,1],firstRun$output[,2,1],ylab="",main="NEP",xlab="Age")
}
