% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_3PG.R
\name{run_3PG}
\alias{run_3PG}
\title{Runs a 3-PG model simulation}
\usage{
run_3PG(site, species, climate, thinning = NULL, parameters = NULL,
  size_dist = NULL, settings = NULL, check_input = TRUE, df_out = TRUE)
}
\arguments{
\item{site}{Data frame as described in \code{\link{prepare_input}} containing site conditions.}

\item{species}{Data frame as described in \code{\link{prepare_input}} containing species-level data.}

\item{climate}{Data frame as described in \code{\link{prepare_input}} containing monthly climatic values.}

\item{thinning}{Data frame as described in \code{\link{prepare_input}} containing thinning information. Default: \code{NULL}.}

\item{parameters}{Data frame as described in \code{\link{prepare_input}} containing parameter values. Default: \code{NULL}.}

\item{size_dist}{Data frame as described in \code{\link{prepare_input}} containing size distribution values. Default: \code{NULL}.}

\item{settings}{List of model settings. See \code{\link{prepare_input}} for details. Default: \code{NULL}.}

\item{check_input}{Logical. If \code{TRUE}, the input will be checked using \code{\link{prepare_input}}. Default: \code{TRUE}.}

\item{df_out}{Logical. If \code{TRUE}, the output will be a long-format data frame. Otherwise, it will be a 4-dimensional array. Default: \code{TRUE}.}
}
\value{
either a 4-dimentional array or a data.frame, depending on the parameter \code{df_out}. More details on the output is \code{\link{i_output}}
}
\description{
Runs the 3-PGpjs (monospecific, even-aged, and evergreen forests) or 3-PGmix (deciduous, uneven-aged, or mixed-species forests) model.
For detailed input requirements, see \code{\link{prepare_input}}.
}
\details{
`r3PG` provides an implementation of the Physiological Processes Predicting Growth \href{https://3pg.forestry.ubc.ca}{3-PG} model, which simulates forest growth and productivity. The `r3PG` serves as a flexible and easy-to-use interface for the `3-PGpjs` (monospecific, evenaged and evergreen forests) and the `3-PGmix` (deciduous, uneven-aged or mixed-species forests) model written in `Fortran`. The package, allows for fast and easy interaction with the model, and `Fortran` re-implementation facilitates computationally intensive sensitivity analysis and calibration. The user can flexibly switch between various options and submodules, to use the original `3-PGpjs` model version for monospecific, even-aged and evergreen forests and the `3-PGmix` model, which can also simulate multi-cohort stands (e.g. mixtures, uneven-aged) that contain deciduous species.

This implementation of 3-PG includes several major variants / modifications of the model in particular the ability to switch between 3-PGpjs (the more classic model version for monospecific stands) vs. 3-PGmix (a version for mixed stands), as well as options for bias corrections and \eqn{\delta^13 C} calculations (see parameters).
}
\note{
The \code{run_3PG} also checks the quality of input data. When names, or structures are not consistent with requirements it will return an error. Turn this off to optimize for speed.
}
\examples{
out <- run_3PG(
  site = d_site,
  species = d_species,
  climate = d_climate,
  thinning = d_thinning,
  parameters = d_parameters,
  size_dist = d_sizeDist,
  settings = list(light_model = 2, transp_model = 2, phys_model = 2,
                  correct_bias = 1, calculate_d13c = 0, mort_model = 1),
  check_input = TRUE, df_out = TRUE) # note that default is TRUE

str(out) # List output format
}
\references{
Forrester, D. I., 2020. 3-PG User Manual. Swiss Federal Institute for Forest, Snow and Landscape Research WSL, Birmensdorf, Switzerland. 70 p. Available at the following web site: \url{http://sites.google.com/site/davidforresterssite/home/projects/3PGmix/3pgmixdownload}

Forrester, D. I., & Tang, X. (2016). Analysing the spatial and temporal dynamics of species interactions in mixed-species forests and the effects of stand density using the 3-PG model. Ecological Modelling, 319, 233–254. \doi{10.1016/j.ecolmodel.2015.07.010}

Landsberg, J. J., & Waring, R. H., 1997. A generalised model of forest productivity using simplified concepts of radiation-use efficiency, carbon balance and partitioning. Forest Ecology and Management, 95(3), 209–228. \doi{10.1016/S0378-1127(97)00026-1}

Sands, P. J., 2010. 3PGpjs user manual. Available at the following web site: \url{https://3pg.sites.olt.ubc.ca/files/2014/04/3PGpjs_UserManual.pdf}
}
\seealso{
\code{\link{prepare_input}}, \code{\link{prepare_parameters}}, \code{\link{prepare_sizeDist}}, \code{\link{prepare_thinning}}, \code{\link{prepare_climate}}
}
