% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_3PG.R
\name{run_3PG}
\alias{run_3PG}
\title{Define the function to run 3PGN in R}
\usage{
run_3PG(
  siteInputs,
  speciesInputs,
  forcingInputs,
  parameterInputs,
  biasInputs,
  settings = list(fit_dbh_dist = 0L)
)
}
\arguments{
\item{siteInputs}{site level data}

\item{speciesInputs}{species level data}

\item{forcingInputs}{forcing level data}

\item{parameterInputs}{parameters level data}

\item{biasInputs}{bial level data}

\item{settings}{settings for the model}
}
\description{
Define the function to run 3PGN in R
}
\details{
This is the model
}
