% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_3PG.R
\name{run_3PG}
\alias{run_3PG}
\title{Define the function to run 3PGN in R}
\usage{
run_3PG(siteInputs, speciesInputs, forcingInputs, parameterInputs,
  biasInputs)
}
\arguments{
\item{siteInputs}{site level data}

\item{speciesInputs}{species level data}

\item{forcingInputs}{forcing level data}

\item{parameterInputs}{parameters level data}

\item{biasInputs}{bial level data}
}
\description{
Define the function to run 3PGN in R
}
\details{
This is the model
}
