% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transf_out.R
\name{transf_out}
\alias{transf_out}
\title{Tranform the output of the fortran 3PG to the long format}
\usage{
transf_out(model)
}
\arguments{
\item{model}{a list as an output of `run_3PG`}
}
\description{
Tranform the output of the fortran 3PG to the long format
}
\details{
This is transforming the model output to long format
}
