% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transf_out.R
\name{transf_out}
\alias{transf_out}
\title{Tranform the output of the fortran 3PG to the long format}
\usage{
transf_out(model)
}
\arguments{
\item{model}{a list as an output of `run_3PG`}
}
\description{
Tranform the output of the fortran 3PG to the long format
}
\details{
This is transforming the model output to long format
}
\examples{
out <- run_3PG(
  siteInputs = site_eum,
  speciesInputs = species_eum,
  forcingInputs = climate_eum,
  managementInputs = NULL,
  parameterInputs = parameters_eum,
  biasInputs = NULL,
  settings = list(light_model = 2, transp_model = 2, phys_model = 2,
                  correct_bias = 0, calculate_d13c = 0))


out_long <- transf_out( out )
}
