% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transf_out.R
\name{transf_out}
\alias{transf_out}
\title{Define the function to run 3PGN in R}
\usage{
transf_out(out, day_start = as.Date("2010-01-31"))
}
\arguments{
\item{out}{the 4 dimentional array as output of `run_3PG`}

\item{day_start}{the first date of the simulation, in date format}
}
\description{
Define the function to run 3PGN in R
}
\details{
This is transforming the model output to long format
}
