% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transf_out.R
\name{transf_out}
\alias{transf_out}
\title{The function to tranform the output to the long format

#' @description This function tranform the output of the \code{\link{run_3PG}} or \code{\link{f_3PG}} functions into the long format dataframe. This allow easy and fast exploration or manipulation of the output data.}
\usage{
transf_out(model)
}
\arguments{
\item{model}{a 2-dimentinal list as an output of \code{\link{run_3PG}} or \code{\link{f_3PG}} function.}
}
\description{
The function to tranform the output to the long format

#' @description This function tranform the output of the \code{\link{run_3PG}} or \code{\link{f_3PG}} functions into the long format dataframe. This allow easy and fast exploration or manipulation of the output data.
}
\details{
Function transform the model simulation from 4-dimentional array to long format dataframe with the following columns: \code{date; species; group; variable; value}.
}
\examples{
out <- run_3PG(
  siteInputs = site_eum,
  speciesInputs = species_eum,
  forcingInputs = climate_eum,
  managementInputs = NULL,
  parameterInputs = parameters_eum,
  biasInputs = bias_eum,
  settings = list(light_model = 2, transp_model = 2, phys_model = 2,
                  correct_bias = 0, calculate_d13c = 0),
  df_out = TRUE)
}
